Fix most of the signal code...
The only problem is that the signal introspection in defClass
gets the internal name instead of the external name... need
to figure out how to deal with this.

Maybe switch the real definition to the external name and then
look at the body of the internal name to figure out what the
real name is?